refactor: #655 interview Session 画面をプレゼンテーション分割する
VAD・MediaRecorder・AudioContext は page に残し、UI のみ SessionScreen へ抽出。
sessionVideoCallbackRef の挙動は維持する。
